requirements.txt
smolagents[openai]
pandas
matplotlib
python-dotenv
